Implement function calls and reference environments using `Rc<Cell<_>>`
# Why do we need `Rc<_>`?
Because of closures. Consider the following case:

    let outer = fn() {
        let a = 1;
        let foo = fn(x) { a + x };
        foo
    }

`foo` needs to have a reference to the environment in `outer`'s body,
i.e., the environment that defines `a`. When outer returns `foo`, this
reference must be mantained, and the environment cannot be dropped.
Only when `foo` eventually goes out of scope can we drop the
environment. Reference counting is one way to solve this. A possible
alternative would be to simply clone the environment when the function
is created, but this has a few issues. Notably, we would run into
problems if `foo` was recursive, because the environment would be closed
before the variable `foo` is assigned, and therefore the environment
owned by the funcion object wouldn't know about `foo`. Furthermore, as
we'll see in the following section, we would also like to be able to
mutate the environment after `foo` is assigned, and ideally we would
like these changes to be reflected in `foo`'s environment.

# Why do we need `Rc<Cell<_>>`?
Consider this slightly modified case:

    let outer = fn() {
        let a = 1;
        let foo = fn(x) { a + x };
        let a = 2;
        foo
    }

If we simply gave the function object an `Rc<Environment>`, it would
necessarily hold shared ownership of the environment, meaning we
couldn't mutate that environment after it was encolosed in `foo`'s
declaration. This is a deal breaker, but using the simpler "cloning the
environment" solution would fix it, right? Well, yes, but (even if we
ignore the issues described earlier), there is some nice behaviour that
we get if we use `Rc<Cell<_>>` to enable mutation. Look closely at the
example. Ideally, we would like the environment to only be closed when
`foo` is returned, meaning `foo` would use the "updated" value of `a`.
This is the behaviour in many programming languages that feature
closures. The simplest (and most performant) way to achieve that
behaviour is to pass around the environment using `Rc<Cell<_>>`,
allowing shared ownership and internal mutability.